Basic start
Just a basic start with some info on my intentions and some of what I expect of others who may contribute back in some fashion in the future when there is actual content.  This will definitely get overhauled with things like Flashing/ Installation instructions and other useful notes as they come in to being.